Refactor: Update SymbolManager to use u32 for symbol_id

+++ src/matching_engine_base.rs
@@ -633,8 +633,8 @@ impl MatchingEngine {
 
 /// Manages symbol-to-ID and ID-to-symbol mappings
 pub struct SymbolManager {
-    pub symbol_to_id: FxHashMap<String, usize>,
-    pub id_to_symbol: FxHashMap<usize, String>,
+    pub symbol_to_id: FxHashMap<String, u32>,
+    pub id_to_symbol: FxHashMap<u32, String>,
 }
 
 impl Default for SymbolManager {
@@ -651,16 +651,16 @@ impl SymbolManager {
         }
     }
 
-    pub fn insert(&mut self, symbol: &str, id: usize) {
+    pub fn insert(&mut self, symbol: &str, id: u32) {
         self.symbol_to_id.insert(symbol.to_string(), id);
         self.id_to_symbol.insert(id, symbol.to_string());
     }
 
-    pub fn get_id(&self, symbol: &str) -> Option<usize> {
+    pub fn get_id(&self, symbol: &str) -> Option<u32> {
         self.symbol_to_id.get(symbol).copied()
     }
 
-    pub fn get_symbol(&self, id: usize) -> Option<&String> {
+    pub fn get_symbol(&self, id: u32) -> Option<&String> {
         self.id_to_symbol.get(&id)
     }
